Feat/logging

+++ Primary/Data/Instrument.cs
@@ -82,11 +82,11 @@ namespace Primary.Data
 
         /// <summary>Minimum volume that can be traded for this instrument.</summary>
         [JsonProperty("minTradeVol")]
-        public uint MinimumTradeVolume { get; set; }
+        public ulong MinimumTradeVolume { get; set; }
 
         /// <summary>Maximum volume that can be traded for this instrument.</summary>
         [JsonProperty("maxTradeVol")]
-        public uint MaximumTradeVolume { get; set; }
+        public ulong MaximumTradeVolume { get; set; }
 
         /// <summary>Lowest price in which it can be traded.</summary>
         [JsonProperty("lowLimitPrice")]
@@ -95,7 +95,6 @@ namespace Primary.Data
         /// <summary>Highest price in which it can be traded.</summary>
         [JsonProperty("highLimitPrice")]
         public decimal MaximumTradePrice { get; set; }
-
     }
 
     public enum InstrumentType
